requirements.txt
lxml==4.8.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
six==1.16.0
XlsxWriter==3.0.3